% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_classification_model.R
\name{butcher_workflow}
\alias{butcher_workflow}
\title{Remove training data from workflow while keeping prediction capability}
\usage{
butcher_workflow(fitted_workflow)
}
\arguments{
\item{fitted_workflow}{A fitted tidymodels workflow}
}
\value{
Butchered workflow with minimal memory footprint
}
\description{
Remove training data from workflow while keeping prediction capability
}
