% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint_utilities.R
\name{cleanup_old_checkpoints}
\alias{cleanup_old_checkpoints}
\title{Remove old checkpoints (keep only most recent N per model type)}
\usage{
cleanup_old_checkpoints(
  checkpoint_dir = "models/classifier/checkpoints",
  keep_n = 2,
  dry_run = TRUE
)
}
\arguments{
\item{checkpoint_dir}{Path to checkpoint directory}

\item{keep_n}{Number of most recent checkpoints to keep per model type}

\item{dry_run}{If TRUE, only show what would be deleted}
}
\value{
Number of files removed
}
\description{
Remove old checkpoints (keep only most recent N per model type)
}
