% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_glmnet.R
\name{create_glmnet_params}
\alias{create_glmnet_params}
\title{Create glmnet parameter definitions for tuning}
\usage{
create_glmnet_params(workflow, train_df)
}
\arguments{
\item{workflow}{Workflow to extract base params from}

\item{train_df}{Training data for sampling ratio bounds}
}
\value{
dials parameter set
}
\description{
Create glmnet parameter definitions for tuning
}
