% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification_evaluation.R
\name{create_lgbm_top_table}
\alias{create_lgbm_top_table}
\title{Create formatted tuning top table for LightGBM}
\usage{
create_lgbm_top_table(topmodels, param_names)
}
\arguments{
\item{topmodels}{Prepared top models}

\item{param_names}{Names of parameters to include in the table}
}
\value{
Formatted top table data frame
}
\description{
Create formatted tuning top table for LightGBM
}
