% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_random_forest.R
\name{create_rf_params}
\alias{create_rf_params}
\title{Create RF parameter definitions for tuning}
\usage{
create_rf_params(workflow, train_df)
}
\arguments{
\item{workflow}{Workflow to extract base params from}

\item{train_df}{Training data}
}
\value{
dials parameter set
}
\description{
Create RF parameter definitions for tuning
}
