% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_bayes_workflow.R
\name{create_sampling_params}
\alias{create_sampling_params}
\title{Create sampling ratio parameter definitions}
\usage{
create_sampling_params(train_df)
}
\arguments{
\item{train_df}{Training data frame with metagenre column}
}
\value{
Named list with under_ratio and over_ratio dials params
}
\description{
Create sampling ratio parameter definitions
}
