% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_bayes_workflow.R
\name{create_tuning_metrics}
\alias{create_tuning_metrics}
\title{Create standard metric set for classification tuning}
\usage{
create_tuning_metrics()
}
\value{
A yardstick metric set
}
\description{
Create standard metric set for classification tuning
}
