% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_classification_model.R
\name{extract_workflow_info}
\alias{extract_workflow_info}
\title{Extract workflow preprocessing steps as documentation}
\usage{
extract_workflow_info(fitted_workflow)
}
\arguments{
\item{fitted_workflow}{Fitted workflow}
}
\value{
List with recipe steps and model spec info
}
\description{
Extract workflow preprocessing steps as documentation
}
