% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenreDiscourseAnalysis-package.R
\docType{package}
\name{GenreDiscourseAnalysis-package}
\alias{GenreDiscourseAnalysis}
\alias{GenreDiscourseAnalysis-package}
\title{GenreDiscourseAnalysis: The GenreDiscourseAnalysis repository provides the necessary functions to analyze and visualize genre discourse}
\description{
The question of the "correct" genre classification of popular music can lead to heated discussions among music fans, music journalists, and musicologists. Nevertheless, empirical studies of music preferences based on self-reported information often assume a uniform understanding among all respondents. In music psychology and music sociology, this assumption often leads to discussions about validity, replication problems, and the inability to compare studies. However, cultural studies research on popular music emphasizes the context- and perspective-dependent attribution of genres as a central phenomenon of everyday and historical discourse on popular music. So why not make a virtue out of necessity? Using genre classifications from different online music communities for 231,370 popular music titles listened to in Germany in recent decades, we present a method for creating a hierarchical genre family tree based on network analysis.
}
\author{
\strong{Maintainer}: Markus Radke \email{markus.radke.1@campus.tu-berlin.de} (\href{https://orcid.org/0009-0001-1719-7060}{ORCID})

}
\keyword{internal}
