% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint_utilities.R
\name{list_checkpoints}
\alias{list_checkpoints}
\title{List all checkpoints in the checkpoint directory}
\usage{
list_checkpoints(checkpoint_dir = "models/classifier/checkpoints")
}
\arguments{
\item{checkpoint_dir}{Path to checkpoint directory}
}
\value{
Data frame with checkpoint information
}
\description{
List all checkpoints in the checkpoint directory
}
