% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genres.R
\name{map_metagenres}
\alias{map_metagenres}
\title{Map tracks to metagenres with probabilities and hard class}
\usage{
map_metagenres(long, metagenre_mapping)
}
\description{
This function is pure and returns a data.frame with one row per track,
probability columns p_\if{html}{\out{<metagenre>}} and a chosen metagenre column.
}
