% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification_evaluation.R
\name{plot_cm}
\alias{plot_cm}
\title{Plot confusion matrix from evaluation object}
\usage{
plot_cm(evaluation, certainty, set = "test")
}
\arguments{
\item{evaluation}{Evaluation object with confusion matrices}

\item{set}{Either "train" or "test"}
}
\description{
Plot confusion matrix from evaluation object
}
