% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_external_contingency.R
\name{plot_external_contingency}
\alias{plot_external_contingency}
\title{Plot contingency between Discogs first-level genres and COMGET metagenres}
\usage{
plot_external_contingency(
  mb_meta,
  poptrag,
  ext_genre_col = "album.dc.firstgenre",
  name = "COMGET-G"
)
}
\arguments{
\item{mb_meta}{A tibble with columns \code{track.s.id} and \code{metagenre}.}

\item{poptrag}{A tibble with columns \code{track.s.id} and external genre col}

\item{external_genre_col}{either  \code{album.dc.firstgenre}, \code{track.dz.album.firstgenre.name}
or \code{track.ab.genrerosamerica}.}
}
\value{
A ggplot2 object (tile plot).
}
\description{
Create a tile plot showing relative frequencies of Discogs first-level
genres (album.dc.firstgenre) within each COMGET metagenre. Tile fill shows
relative frequency and tile labels show absolute counts.
}
