% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_classification_evaluation_plots.R
\name{plot_parameter_tuning}
\alias{plot_parameter_tuning}
\title{Plot parameter tuning exploration}
\usage{
plot_parameter_tuning(tuning_history, parameter, show_legend = FALSE)
}
\arguments{
\item{tuning_history}{Tuning history data frame}

\item{parameter}{Parameter name to plot}

\item{show_legend}{Whether to show the legend}
}
\description{
Plot parameter tuning exploration
}
