% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification_evaluation.R
\name{plot_tuning_model_selection}
\alias{plot_tuning_model_selection}
\title{Plot tuning model selection graph}
\usage{
plot_tuning_model_selection(topmodels, top_n)
}
\arguments{
\item{topmodels}{Prepared top models from prepare_tuning_topmodels}

\item{top_n}{Number of models shown}
}
\description{
Plot tuning model selection graph
}
