% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_classification_evaluation_plots.R
\name{prepare_tuning_topmodels}
\alias{prepare_tuning_topmodels}
\title{Prepare tuning top models with ranking and model grouping}
\usage{
prepare_tuning_topmodels(tuning_history, chosen_params, top_n)
}
\arguments{
\item{tuning_history}{Tuning history data frame}

\item{chosen_params}{Named vector of chosen parameter values}

\item{top_n}{Number of top models to return}
}
\description{
Prepare tuning top models with ranking and model grouping
}
