% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint_utilities.R
\name{remove_checkpoints_by_hash}
\alias{remove_checkpoints_by_hash}
\title{Remove checkpoints for a specific model hash}
\usage{
remove_checkpoints_by_hash(
  checkpoint_dir = "models/classifier/checkpoints",
  model_hash,
  dry_run = TRUE
)
}
\arguments{
\item{checkpoint_dir}{Path to checkpoint directory}

\item{model_hash}{Hash of the model (8 characters or full)}

\item{dry_run}{If TRUE, only show what would be deleted}
}
\value{
Number of files removed
}
\description{
Remove checkpoints for a specific model hash
}
