% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_classification_evaluation_plots.R
\name{render_glmnet_varimp_table}
\alias{render_glmnet_varimp_table}
\title{Render GLMNET variable importance table}
\usage{
render_glmnet_varimp_table(model, n_top = 5, caption = NULL)
}
\arguments{
\item{model}{GLMNET model object}

\item{n_top}{Number of top features per class}
}
\description{
Render GLMNET variable importance table
}
