% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classification_evaluation.R
\name{render_tuning_top_table}
\alias{render_tuning_top_table}
\title{Render tuning top table with kable styling}
\usage{
render_tuning_top_table(top_table)
}
\arguments{
\item{top_table}{Formatted table from create_*_top_table}

\item{col_names}{Column names for the table}
}
\description{
Render tuning top table with kable styling
}
