% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_adaptive_sampling.R
\name{target_ratio}
\alias{target_ratio}
\title{Create a target_ratio dials parameter}
\usage{
target_ratio(range = c(1, 20), trans = NULL)
}
\arguments{
\item{range}{Numeric vector of length 2 with min and max values}

\item{trans}{Transformation to apply (NULL for identity)}
}
\value{
A dials parameter object
}
\description{
Create a target_ratio dials parameter
}
