% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_random_forest.R
\name{train_random_forest}
\alias{train_random_forest}
\title{Train and evaluate random-forest models using tidymodels}
\usage{
train_random_forest(train, test, cv_splits, settings)
}
\arguments{
\item{settings}{Named list containing \code{model_features}, \code{under_ratio_fix}, \code{over_ratio_fix},
\code{max_depth_fix}, \code{tune_downsample}, \code{tune_upsample},
\code{varimp_top_n}, \code{seed},  \code{tune_hyperparameters}, \code{use_caseweights},
\code{n_cores}, and optionally \code{ntrees}, \code{mtry_fix}, \code{min.node.size_fix},
\code{importance}, \code{initial_grid_density}, \code{bayes_iterations}, \code{uncertain_jump}.}

\item{datasets}{List containing \code{train}, \code{test},
and optionally \code{cv_splits} (required for tuning).}
}
\value{
List with \code{low} and \code{high} results containing \code{model}, \code{model_settings},
\code{evaluation}, \code{train_df}, \code{test_df}, and optionally \code{tuning_results}.
}
\description{
Train and evaluate random-forest models using tidymodels
}
